Fold the new tests into the existing per-source test files
The previous commits added parallel test files next to the ones that
were already there, so a package ended up with both node-repo_test.go
and node-repo-datastore_test.go covering the same source. Merged each
of them into the file that already covers that source and deleted the
extras:

  node-lifecycle -> core/node/node_test.go
  gossip-topics  -> core/pubsub/gossip_test.go
  stream-pool    -> core/stream/stream_test.go
  db-lifecycle   -> database/local-store/db_test.go
  follower-graph -> database/follower-repo_test.go
  node-datastore -> database/node-repo_test.go
  tweet-lifecycle-> database/tweet-repo_test.go
  aes-codec      -> security/aes_test.go

Three files covered more than one source and were split accordingly:
media -> image_test.go + video_test.go, social -> block_test.go +
following_test.go + engagement_test.go, media-user -> media-repo_test.go
+ user-repo_test.go, with the media and user cases folded into the
suites those files already define instead of introducing new ones.

Also widened the margin in TestPutWithTTLAndSetTTL: a 2s TTL with a 1s
sleep left one second for the write, the sleep and SetTTL, and the key
expired under the load of the full suite.

No test bodies changed; coverage is unchanged at 72.0% (CI metric
70.7%).
