+++ core/node/node_test.go
@@ -26,11 +26,22 @@ package node
 
 import (
 	"context"
+	"crypto/ed25519"
+	"crypto/rand"
+	"errors"
 	"fmt"
-	"io"
-	"testing"
-
+	"github.com/Warp-net/warpnet/core/stream"
 	"github.com/Warp-net/warpnet/core/warpnet"
+	warpevent "github.com/Warp-net/warpnet/event"
+	"github.com/Warp-net/warpnet/json"
+	"github.com/Warp-net/warpnet/security"
+	"github.com/libp2p/go-libp2p"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"io"
+	"sync"
+	"testing"
+	"time"
 )
 
 func TestIsBenignStreamCloseErr(t *testing.T) {
@@ -56,4 +67,520 @@ func TestIsBenignStreamCloseErr(t *testing.T) {
 			}
 		})
 	}
+}
+
+const echoRoute = stream.WarpRoute("/public/get/user/0.0.0")
+
+func newTestNode(t *testing.T) *WarpNode {
+	t.Helper()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	t.Cleanup(cancel)
+
+	n, err := NewWarpNode(ctx, libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"))
+	require.NoError(t, err)
+	t.Cleanup(func() {
+		defer func() { _ = recover() }()
+		n.StopNode()
+	})
+	return n
+}
+
+// signedEnvelope builds the wire message a self-stream expects: a full
+// event.Message signed with this node's own key. Anything less is rejected by
+// the auth middleware, which is exactly what the loopback path must enforce.
+func signedEnvelope(t *testing.T, n *WarpNode, path stream.WarpRoute, body []byte) []byte {
+	t.Helper()
+
+	priv, err := n.Node().Peerstore().PrivKey(n.Node().ID()).Raw()
+	require.NoError(t, err)
+
+	msg := warpevent.Message{
+		Body:        body,
+		MessageId:   "self-" + n.Node().ID().String(),
+		NodeId:      n.Node().ID().String(),
+		Destination: string(path),
+		Timestamp:   time.Now().UTC(),
+		Version:     "0.0.0",
+	}
+	msg.Signature = security.Sign(priv, msg.SigningBytes())
+
+	bt, err := json.Marshal(msg)
+	require.NoError(t, err)
+	return bt
+}
+
+// ---------------------------------------------------------------------------
+// Construction and identity.
+// ---------------------------------------------------------------------------
+
+func TestWarpNode_StartsAndReportsItself(t *testing.T) {
+	n := newTestNode(t)
+
+	require.NotNil(t, n.Node())
+	assert.NotEmpty(t, n.Node().ID().String())
+
+	info := n.BaseNodeInfo()
+	assert.Equal(t, n.Node().ID(), info.ID)
+	assert.NotEmpty(t, info.Addresses, "a listening node must advertise at least one address")
+	assert.False(t, info.StartTime.IsZero())
+	assert.NotNil(t, n.Prioritizer())
+}
+
+func TestWarpNode_NilReceiverIsInert(t *testing.T) {
+	var n *WarpNode
+
+	assert.Equal(t, warpnet.NodeInfo{}, n.BaseNodeInfo())
+	assert.Nil(t, n.Node())
+	assert.NoError(t, n.Connect(warpnet.WarpAddrInfo{}))
+
+	_, err := n.Stream("peer", echoRoute, nil)
+	assert.Error(t, err, "an uninitialised node must report, not panic")
+
+	assert.NotPanics(t, func() { n.StopNode() })
+	assert.NotPanics(t, func() { n.SetOutbox(nil) })
+}
+
+// A node identity must be derived from the supplied key, so a restart with the
+// same seed keeps the same peer id and followers stay reachable.
+func TestWarpIdentity_IsDeterministicAndSingular(t *testing.T) {
+	_, priv, err := ed25519.GenerateKey(rand.Reader)
+	require.NoError(t, err)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	first, err := NewWarpNode(ctx, WarpIdentity(priv), libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"))
+	require.NoError(t, err)
+	id := first.Node().ID()
+	first.StopNode()
+
+	second, err := NewWarpNode(ctx, WarpIdentity(priv), libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"))
+	require.NoError(t, err)
+	defer second.StopNode()
+
+	assert.Equal(t, id, second.Node().ID(), "the same key must yield the same peer id")
+
+	// Supplying two identities is a configuration error, not a silent pick.
+	_, err = NewWarpNode(ctx, WarpIdentity(priv), WarpIdentity(priv))
+	assert.ErrorIs(t, err, ErrMultipleIdentities)
+}
+
+func TestWarpIdentity_PanicsOnMalformedKey(t *testing.T) {
+	assert.Panics(t, func() { WarpIdentity(make(ed25519.PrivateKey, 7)) },
+		"a malformed node key must fail loudly at startup, not produce a random identity")
+}
+
+// ---------------------------------------------------------------------------
+// Handler registration.
+// ---------------------------------------------------------------------------
+
+// An invalid route must abort startup — a silently unregistered handler makes
+// the feature look broken at runtime with no clue why.
+func TestSetStreamHandlers_PanicsOnInvalidRoute(t *testing.T) {
+	n := newTestNode(t)
+
+	assert.Panics(t, func() {
+		n.SetStreamHandlers(warpnet.WarpStreamHandler{
+			Path:    warpnet.WarpProtocolID("garbage-route"),
+			Handler: func([]byte, warpnet.WarpStream) (any, error) { return nil, nil },
+		})
+	})
+}
+
+func TestSetStreamHandlers_RegistersOnTheHost(t *testing.T) {
+	n := newTestNode(t)
+
+	n.SetStreamHandlers(warpnet.WarpStreamHandler{
+		Path:    warpnet.WarpProtocolID(echoRoute),
+		Handler: func([]byte, warpnet.WarpStream) (any, error) { return []byte(`{}`), nil },
+	})
+
+	assert.Contains(t, n.Node().Mux().Protocols(), warpnet.WarpProtocolID(echoRoute))
+	assert.Contains(t, n.BaseNodeInfo().Protocols, warpnet.WarpProtocolID(echoRoute))
+}
+
+// ---------------------------------------------------------------------------
+// SelfStream — the loopback path gossip uses to replay messages locally.
+// ---------------------------------------------------------------------------
+
+func TestSelfStream_RoundTripsThroughTheRegisteredHandler(t *testing.T) {
+	n := newTestNode(t)
+
+	var seen []byte
+	n.SetStreamHandlers(warpnet.WarpStreamHandler{
+		Path: warpnet.WarpProtocolID(echoRoute),
+		Handler: func(msg []byte, s warpnet.WarpStream) (any, error) {
+			seen = msg
+			return []byte(`{"pong":true}`), nil
+		},
+	})
+
+	resp, err := n.SelfStream(echoRoute, signedEnvelope(t, n, echoRoute, []byte(`{"ping":true}`)))
+	require.NoError(t, err)
+	assert.Equal(t, `{"pong":true}`, string(resp))
+	assert.Equal(t, `{"ping":true}`, string(seen), "the handler must see the message body, not the envelope")
+}
+
+// An unsigned or forged self-stream must be refused just like a remote one —
+// the loopback path is not a way around authentication.
+func TestSelfStream_RejectsUnsignedAndForgedEnvelopes(t *testing.T) {
+	n := newTestNode(t)
+
+	var handlerCalls int
+	n.SetStreamHandlers(warpnet.WarpStreamHandler{
+		Path: warpnet.WarpProtocolID(echoRoute),
+		Handler: func([]byte, warpnet.WarpStream) (any, error) {
+			handlerCalls++
+			return []byte(`{"ok":true}`), nil
+		},
+	})
+
+	// A bare body with no envelope at all.
+	resp, err := n.SelfStream(echoRoute, []byte(`{"ping":true}`))
+	require.NoError(t, err)
+	assert.NotContains(t, string(resp), `"ok":true`)
+
+	// A well-formed envelope with no signature.
+	unsigned, err := json.Marshal(warpevent.Message{
+		Body:      []byte(`{}`),
+		MessageId: "no-signature",
+		Timestamp: time.Now().UTC(),
+	})
+	require.NoError(t, err)
+	resp, err = n.SelfStream(echoRoute, unsigned)
+	require.NoError(t, err)
+	assert.NotContains(t, string(resp), `"ok":true`)
+
+	// A signature that does not match the body.
+	forged := signedEnvelope(t, n, echoRoute, []byte(`{"original":true}`))
+	var msg warpevent.Message
+	require.NoError(t, json.Unmarshal(forged, &msg))
+	msg.Body = []byte(`{"tampered":true}`)
+	tampered, err := json.Marshal(msg)
+	require.NoError(t, err)
+
+	resp, err = n.SelfStream(echoRoute, tampered)
+	require.NoError(t, err)
+	assert.NotContains(t, string(resp), `"ok":true`)
+
+	assert.Zero(t, handlerCalls, "no unauthenticated message may reach a handler")
+}
+
+func TestSelfStream_RejectsEmptyDataAndUnknownRoute(t *testing.T) {
+	n := newTestNode(t)
+
+	_, err := n.SelfStream(echoRoute, nil)
+	assert.Error(t, err, "an empty self-stream is a programming error")
+
+	_, err = n.SelfStream(stream.WarpRoute("/public/get/nothing/0.0.0"), []byte(`{}`))
+	assert.Error(t, err, "an unregistered route must be reported, not silently dropped")
+}
+
+// A handler that fails must still produce an answer rather than hanging the
+// caller waiting for bytes.
+func TestSelfStream_HandlerErrorStillAnswers(t *testing.T) {
+	n := newTestNode(t)
+
+	n.SetStreamHandlers(warpnet.WarpStreamHandler{
+		Path: warpnet.WarpProtocolID(echoRoute),
+		Handler: func([]byte, warpnet.WarpStream) (any, error) {
+			return nil, errors.New("handler exploded")
+		},
+	})
+
+	envelope := signedEnvelope(t, n, echoRoute, []byte(`{}`))
+
+	done := make(chan struct{})
+	var resp []byte
+	go func() {
+		defer close(done)
+		resp, _ = n.SelfStream(echoRoute, envelope)
+	}()
+
+	select {
+	case <-done:
+	case <-time.After(30 * time.Second):
+		t.Fatal("a failing self-stream handler hung the caller")
+	}
+	assert.Contains(t, string(resp), "handler exploded")
+}
+
+// ---------------------------------------------------------------------------
+// Remote streaming.
+// ---------------------------------------------------------------------------
+
+// A node must never open a network stream to itself: that is a guaranteed
+// deadlock against its own handler.
+func TestStream_RefusesSelfRequest(t *testing.T) {
+	n := newTestNode(t)
+
+	_, err := n.Stream(n.Node().ID(), echoRoute, []byte(`{}`))
+	assert.ErrorIs(t, err, ErrSelfRequest)
+}
+
+func TestStream_ReachesAnotherNode(t *testing.T) {
+	server := newTestNode(t)
+	client := newTestNode(t)
+
+	server.SetStreamHandlers(warpnet.WarpStreamHandler{
+		Path: warpnet.WarpProtocolID(echoRoute),
+		Handler: func(msg []byte, s warpnet.WarpStream) (any, error) {
+			return []byte(`{"from":"server"}`), nil
+		},
+	})
+
+	require.NoError(t, client.Connect(warpnet.WarpAddrInfo{
+		ID: server.Node().ID(), Addrs: server.Node().Addrs(),
+	}))
+
+	resp, err := client.Stream(server.Node().ID(), echoRoute, []byte(`{"hi":true}`))
+	require.NoError(t, err)
+	assert.Contains(t, string(resp), `"from":"server"`)
+}
+
+func TestStream_UnknownPeerIsOffline(t *testing.T) {
+	n := newTestNode(t)
+
+	stranger := warpnet.FromStringToPeerID("12D3KooWSjbYrsVoXzJcEtmgJLMVCbPXMzJmNN1JkEZB9LJ2rnmU")
+	require.NotEmpty(t, string(stranger))
+
+	_, err := n.Stream(stranger, echoRoute, []byte(`{}`))
+	assert.Error(t, err, "a peer we have never heard of cannot be streamed to")
+}
+
+func TestConnect_IsIdempotentAndIgnoresEmptyPeer(t *testing.T) {
+	server := newTestNode(t)
+	client := newTestNode(t)
+
+	info := warpnet.WarpAddrInfo{ID: server.Node().ID(), Addrs: server.Node().Addrs()}
+
+	require.NoError(t, client.Connect(info))
+	assert.NoError(t, client.Connect(info), "re-connecting an already-connected peer is a no-op")
+}
+
+func TestStopNode_IsSafeToCallTwice(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	n, err := NewWarpNode(ctx, libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"))
+	require.NoError(t, err)
+
+	n.StopNode()
+	assert.NotPanics(t, func() { n.StopNode() }, "shutdown must be idempotent")
+}
+
+// ---------------------------------------------------------------------------
+// Reachability prioritisation.
+// ---------------------------------------------------------------------------
+
+type recordingConnManager struct {
+	warpnet.WarpConnManager
+
+	mu   sync.Mutex
+	tags map[string]int
+}
+
+func newRecordingConnManager() *recordingConnManager {
+	return &recordingConnManager{tags: map[string]int{}}
+}
+
+func (m *recordingConnManager) UpsertTag(p warpnet.WarpPeerID, tag string, upsert func(int) int) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
+	m.tags[p.String()] = upsert(m.tags[p.String()])
+}
+
+func (m *recordingConnManager) get(p warpnet.WarpPeerID) (int, bool) {
+	m.mu.Lock()
+	defer m.mu.Unlock()
+	v, ok := m.tags[p.String()]
+	return v, ok
+}
+
+// A publicly reachable peer is more valuable to keep connected than a private
+// one — the connection manager trims by this weight under pressure.
+func TestReachabilityManager_RanksByReachability(t *testing.T) {
+	cases := []struct {
+		reach warpnet.WarpReachability
+		want  int
+	}{
+		{warpnet.ReachabilityPublic, 90},
+		{warpnet.ReachabilityUnknown, 60},
+		{warpnet.ReachabilityPrivate, 30},
+	}
+
+	for i, c := range cases {
+		cm := newRecordingConnManager()
+		m := newNodeReachabilityManager(cm)
+		pid := warpnet.FromStringToPeerID("12D3KooWSjbYrsVoXzJcEtmgJLMVCbPXMzJmNN1JkEZB9LJ2rnmU")
+
+		m.SetPriority(pid, c.reach)
+
+		got, ok := cm.get(pid)
+		require.Truef(t, ok, "case %d: no tag recorded", i)
+		assert.Equalf(t, c.want, got, "case %d", i)
+	}
+}
+
+func TestReachabilityManager_MinAndMaxPriority(t *testing.T) {
+	pid := warpnet.FromStringToPeerID("12D3KooWSjbYrsVoXzJcEtmgJLMVCbPXMzJmNN1JkEZB9LJ2rnmU")
+
+	cm := newRecordingConnManager()
+	newNodeReachabilityManager(cm).SetMinPriority(pid)
+	got, ok := cm.get(pid)
+	require.True(t, ok)
+	assert.Equal(t, 1, got, "a min-priority peer is the first to be trimmed")
+
+	cm = newRecordingConnManager()
+	newNodeReachabilityManager(cm).SetMaxPriority(pid)
+	got, ok = cm.get(pid)
+	require.True(t, ok)
+	assert.Equal(t, 100, got, "a relay must be the last connection dropped")
+}
+
+// A peer whose reachability flaps must not be re-tagged on every event —
+// otherwise the connection manager churns instead of stabilising.
+func TestReachabilityManager_SuppressesFlapping(t *testing.T) {
+	cm := newRecordingConnManager()
+	m := newNodeReachabilityManager(cm)
+	pid := warpnet.FromStringToPeerID("12D3KooWSjbYrsVoXzJcEtmgJLMVCbPXMzJmNN1JkEZB9LJ2rnmU")
+
+	m.SetPriority(pid, warpnet.ReachabilityPublic)
+	first, ok := cm.get(pid)
+	require.True(t, ok)
+
+	// Immediately contradicting itself must be ignored while the mark is warm.
+	m.SetPriority(pid, warpnet.ReachabilityPrivate)
+	m.SetMinPriority(pid)
+	m.SetMaxPriority(pid)
+
+	second, _ := cm.get(pid)
+	assert.Equal(t, first, second, "a flapping peer must keep its first verdict")
+}
+
+// ---------------------------------------------------------------------------
+// libp2p option helpers.
+// ---------------------------------------------------------------------------
+
+// A node must never list itself as its own relay candidate.
+func TestEnableAutoRelayWithStaticRelays_DropsSelf(t *testing.T) {
+	self := warpnet.FromStringToPeerID("12D3KooWMKZFrp1BDKg9amtkv5zWnLhuUXN32nhqMvbtMdV2hz7j")
+	other := warpnet.FromStringToPeerID("12D3KooWSjbYrsVoXzJcEtmgJLMVCbPXMzJmNN1JkEZB9LJ2rnmU")
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	opt := EnableAutoRelayWithStaticRelays(
+		[]warpnet.WarpAddrInfo{{ID: self}, {ID: other}}, self,
+	)
+	n, err := NewWarpNode(ctx, opt(), libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"))
+	require.NoError(t, err)
+	defer n.StopNode()
+
+	assert.NotEmpty(t, n.Node().ID().String())
+}
+
+// With no relays left after filtering, auto-relay must simply stay off rather
+// than failing startup.
+func TestEnableAutoRelayWithStaticRelays_EmptyListIsInert(t *testing.T) {
+	self := warpnet.FromStringToPeerID("12D3KooWMKZFrp1BDKg9amtkv5zWnLhuUXN32nhqMvbtMdV2hz7j")
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	opt := EnableAutoRelayWithStaticRelays([]warpnet.WarpAddrInfo{{ID: self}}, self)
+	n, err := NewWarpNode(ctx, opt(), libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"))
+	require.NoError(t, err)
+	defer n.StopNode()
+}
+
+func TestEmptyOption_IsANoOp(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	n, err := NewWarpNode(ctx, EmptyOption()(), libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"))
+	require.NoError(t, err)
+	defer n.StopNode()
+	assert.NotNil(t, n.Node())
+}
+
+// The dial timeouts are set through reflection into private libp2p fields; if
+// upstream renames them the option must report it instead of silently leaving
+// the default in place.
+func TestPrivateFieldOptionsStillMatchUpstream(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	n, err := NewWarpNode(ctx,
+		libp2p.SwarmOpts(WithDialTimeout(3*time.Second), WithDialTimeoutLocal(2*time.Second)),
+		libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"),
+	)
+	require.NoError(t, err, "a renamed libp2p field must not break node startup")
+	defer n.StopNode()
+
+	sw, ok := n.Node().Network().(*warpnet.Swarm)
+	require.True(t, ok)
+
+	assert.NoError(t, WithDialTimeout(time.Second)(sw))
+	assert.NoError(t, WithDialTimeoutLocal(time.Second)(sw))
+}
+
+func TestSetPrivateDurationField_ReportsMissingField(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	n, err := NewWarpNode(ctx, libp2p.ListenAddrStrings("/ip4/127.0.0.1/tcp/0"))
+	require.NoError(t, err)
+	defer n.StopNode()
+
+	sw, ok := n.Node().Network().(*warpnet.Swarm)
+	require.True(t, ok)
+
+	err = setPrivateDurationField(sw, "thisFieldDoesNotExist", time.Second)
+	assert.ErrorIs(t, err, ErrFieldNotFound)
+}
+
+func TestWithDefaultTCPConnectionTimeout_ReportsMissingField(t *testing.T) {
+	tr := &warpnet.TCPTransport{}
+	assert.NoError(t, WithDefaultTCPConnectionTimeout(5*time.Second)(tr))
+}
+
+// ---------------------------------------------------------------------------
+// Tracers must survive malformed events rather than taking the node down.
+// ---------------------------------------------------------------------------
+
+func TestHolePunchTracer_ToleratesNilAndUnknownEvents(t *testing.T) {
+	tr := holePunchTracer{}
+	assert.NotPanics(t, func() { tr.Trace(nil) })
+}
+
+func TestConnTracer_ToleratesNilConnections(t *testing.T) {
+	tr := connTracer{}
+	assert.NotPanics(t, func() {
+		tr.Listen(nil, nil)
+		tr.ListenClose(nil, nil)
+		tr.Connected(nil, nil)
+		tr.Disconnected(nil, nil)
+	})
+}
+
+func TestConnTracer_ClassifiesRealConnections(t *testing.T) {
+	server := newTestNode(t)
+	client := newTestNode(t)
+
+	require.NoError(t, client.Connect(warpnet.WarpAddrInfo{
+		ID: server.Node().ID(), Addrs: server.Node().Addrs(),
+	}))
+
+	conns := client.Node().Network().ConnsToPeer(server.Node().ID())
+	require.NotEmpty(t, conns)
+
+	assert.False(t, isRelayed(conns[0]), "a direct TCP dial is not relayed")
+
+	tr := connTracer{}
+	assert.NotPanics(t, func() {
+		tr.Connected(client.Node().Network(), conns[0])
+		tr.Disconnected(client.Node().Network(), conns[0])
+	})
 }
