+++ security/aes_test.go
@@ -30,9 +30,10 @@ package security
 
 import (
 	"bytes"
-	"testing"
-
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"strings"
+	"testing"
 )
 
 func TestAESEncryptDecrypt_Success(t *testing.T) {
@@ -104,5 +105,180 @@ func TestAESDecrypt_TooShort(t *testing.T) {
 	assert.ErrorIs(t, err, ErrCiphertextTooShort)
 
 	_, err = decryptAES(make([]byte, saltSize+nonceSize-1), []byte("pw"))
+	assert.ErrorIs(t, err, ErrCiphertextTooShort)
+}
+func TestAESKeyFromPassword_IsStableAndDistinct(t *testing.T) {
+	a := AESKeyFromPassword("hunter2")
+	b := AESKeyFromPassword("hunter2")
+	c := AESKeyFromPassword("hunter3")
+
+	assert.Len(t, a, 32, "AES-256 needs a 32-byte key")
+	assert.Equal(t, a, b, "the same dashboard password must derive the same key")
+	assert.NotEqual(t, a, c, "a different password must not collide")
+
+	// An empty password still yields a usable key rather than a nil slice that
+	// would silently disable encryption.
+	assert.Len(t, AESKeyFromPassword(""), 32)
+}
+
+func TestAESCodec_RoundTrip(t *testing.T) {
+	codec := AESCodec{Key: AESKeyFromPassword("dashboard-secret")}
+
+	plain := []byte(`{"text":"привет 🔥","private":true}`)
+
+	sealed, err := codec.Encode(plain, true)
+	require.NoError(t, err)
+	assert.NotEqual(t, plain, sealed, "the payload must not travel in the clear")
+	assert.NotContains(t, string(sealed), "привет")
+
+	got, encrypted := codec.Decode(sealed)
+	assert.True(t, encrypted)
+	assert.Equal(t, plain, got)
+}
+
+// Two encryptions of the same plaintext must differ — a deterministic nonce
+// would leak that a user re-sent the same message.
+func TestAESCodec_NonceIsFresh(t *testing.T) {
+	codec := AESCodec{Key: AESKeyFromPassword("secret")}
+	plain := []byte("same message")
+
+	first, err := codec.Encode(plain, true)
+	require.NoError(t, err)
+	second, err := codec.Encode(plain, true)
+	require.NoError(t, err)
+
+	assert.NotEqual(t, first, second, "ciphertexts must not repeat for identical input")
+
+	one, ok := codec.Decode(first)
+	assert.True(t, ok)
+	two, ok := codec.Decode(second)
+	assert.True(t, ok)
+	assert.Equal(t, one, two)
+}
+
+// A codec with no key is the "no dashboard password" configuration: traffic
+// passes through untouched and is honestly reported as unencrypted.
+func TestAESCodec_NoKeyIsPassthrough(t *testing.T) {
+	codec := AESCodec{}
+	payload := []byte("plain text")
+
+	out, err := codec.Encode(payload, true)
+	require.NoError(t, err)
+	assert.Equal(t, payload, out)
+
+	got, encrypted := codec.Decode(payload)
+	assert.Equal(t, payload, got)
+	assert.False(t, encrypted, "an unkeyed codec must never claim a frame was encrypted")
+}
+
+func TestAESCodec_EncodeRespectsEncryptedFlag(t *testing.T) {
+	codec := AESCodec{Key: AESKeyFromPassword("secret")}
+	payload := []byte("public notice")
+
+	out, err := codec.Encode(payload, false)
+	require.NoError(t, err)
+	assert.Equal(t, payload, out, "a frame marked plain must go out plain")
+}
+
+// The wrong password must not decrypt someone else's dashboard traffic. The
+// codec degrades to "plaintext" rather than returning garbage as if it were
+// authentic.
+func TestAESCodec_WrongKeyDoesNotDecrypt(t *testing.T) {
+	sender := AESCodec{Key: AESKeyFromPassword("correct-password")}
+	attacker := AESCodec{Key: AESKeyFromPassword("guessed-password")}
+
+	plain := []byte("session token: abc123")
+	sealed, err := sender.Encode(plain, true)
+	require.NoError(t, err)
+
+	got, encrypted := attacker.Decode(sealed)
+	assert.False(t, encrypted, "a wrong key must never report a successful decrypt")
+	assert.NotEqual(t, plain, got)
+	assert.Equal(t, sealed, got, "the frame is handed back untouched")
+}
+
+// GCM is authenticated: flipping a byte must fail the tag, not yield a mangled
+// but accepted message.
+func TestAESCodec_TamperedCiphertextIsRejected(t *testing.T) {
+	codec := AESCodec{Key: AESKeyFromPassword("secret")}
+
+	sealed, err := codec.Encode([]byte("transfer 10 to alice"), true)
+	require.NoError(t, err)
+
+	tampered := bytes.Clone(sealed)
+	// Flip a character in the middle of the base64 body.
+	idx := len(tampered) / 2
+	if tampered[idx] == 'A' {
+		tampered[idx] = 'B'
+	} else {
+		tampered[idx] = 'A'
+	}
+
+	got, encrypted := codec.Decode(tampered)
+	assert.False(t, encrypted, "a tampered frame must not authenticate")
+	assert.Equal(t, tampered, got)
+}
+
+func TestAESCodec_MalformedFramesAreRejectedNotPanicking(t *testing.T) {
+	codec := AESCodec{Key: AESKeyFromPassword("secret")}
+
+	frames := [][]byte{
+		nil,
+		{},
+		[]byte("not base64 at all !!!"),
+		[]byte("AAAA"),                       // valid base64, far too short for a nonce
+		[]byte(strings.Repeat("A", 16)),      // still shorter than nonce+tag
+		[]byte("////"),                       // valid base64, junk bytes
+		[]byte(strings.Repeat("QUJD", 1000)), // long but meaningless
+	}
+
+	for _, frame := range frames {
+		assert.NotPanics(t, func() {
+			got, encrypted := codec.Decode(frame)
+			assert.False(t, encrypted)
+			assert.Equal(t, frame, got)
+		})
+	}
+}
+
+func TestAESCodec_EmptyPayloadRoundTrips(t *testing.T) {
+	codec := AESCodec{Key: AESKeyFromPassword("secret")}
+
+	sealed, err := codec.Encode([]byte{}, true)
+	require.NoError(t, err)
+	assert.NotEmpty(t, sealed, "even an empty body carries a nonce and tag")
+
+	got, encrypted := codec.Decode(sealed)
+	assert.True(t, encrypted)
+	assert.Empty(t, got)
+}
+
+func TestAESCodec_LargePayloadRoundTrips(t *testing.T) {
+	codec := AESCodec{Key: AESKeyFromPassword("secret")}
+
+	plain := bytes.Repeat([]byte("warpnet"), 100_000)
+	sealed, err := codec.Encode(plain, true)
+	require.NoError(t, err)
+
+	got, encrypted := codec.Decode(sealed)
+	assert.True(t, encrypted)
+	assert.Equal(t, plain, got)
+}
+
+func TestAESGCM_RejectsInvalidKeySizes(t *testing.T) {
+	for _, size := range []int{0, 1, 15, 17, 31, 33, 64} {
+		_, err := aesGCMEncrypt(make([]byte, size), []byte("x"))
+		assert.Errorf(t, err, "key size %d must be rejected", size)
+
+		_, err = aesGCMDecrypt(make([]byte, size), []byte("QUJD"))
+		assert.Errorf(t, err, "key size %d must be rejected", size)
+	}
+}
+
+func TestAESGCM_ShortCiphertextIsReported(t *testing.T) {
+	key := AESKeyFromPassword("secret")
+
+	// Valid base64 that decodes to fewer bytes than the nonce.
+	_, err := aesGCMDecrypt(key, []byte("QUJD"))
 	assert.ErrorIs(t, err, ErrCiphertextTooShort)
 }
